backend/app/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "main application"
authors = [
    {name = "ZeroRangee"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "asgiref (>=3.9.1,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg[pool] (>=3.2.9,<4.0.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


[tool.poetry.scripts]
start = "start:start"